REPLY UPDATE AJAX 완성

+++ src/main/java/kr/co/bttf/domain/ReplyVO.java
@@ -6,10 +6,8 @@ import lombok.Data;
 
 @Data
 public class ReplyVO {
-	private Long reply_id;
-	private Long post_id;
-	//private int reply_depth;
-	//private int bundle_id;
+	private int reply_id;
+	private int post_id;
 	private String reply_contents;
 	private String user_nickname;
 	private Date reply_regdate;
